% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createPlot}
\alias{createPlot}
\title{Create Plot Function}
\usage{
createPlot(plot_type_value, var_plot_value, data_value)
}
\arguments{
\item{plot_type_value}{Character string specifying the type of plot to create.
Options include: "Bar plot", "Box plot", "Correlation plot", "Density plot",
"Dot plot", "Histogram", "QQ plot", "Scatter and Line plot".}

\item{var_plot_value}{Character vector of variable names to plot.}

\item{data_value}{Data frame containing the data to plot.}
}
\value{
A list of ggplot objects or a single ggplot object depending on the plot type.
}
\description{
Creates various types of plots for data visualization in the StepReg Shiny application.
}
